+++ src/main/java/insane96mcp/insanelib/InsaneLib.java
@@ -1,5 +1,6 @@
 package insane96mcp.insanelib;
 
+import insane96mcp.insanelib.module.Modules;
 import insane96mcp.insanelib.setup.ClientSetup;
 import insane96mcp.insanelib.setup.Config;
 import insane96mcp.insanelib.setup.ILEntities;
@@ -18,10 +19,10 @@ public class InsaneLib
     public static final String RESOURCE_PREFIX = MOD_ID + ":";
 
     public InsaneLib() {
+        Modules.init();
         ModLoadingContext.get().registerConfig(net.minecraftforge.fml.config.ModConfig.Type.COMMON, Config.COMMON_SPEC);
         FMLJavaModLoadingContext.get().getModEventBus().addListener(ClientSetup::init);
         final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
         ILEntities.ENTITIES.register(modEventBus);
-        //Module.loadFeatures(MOD_ID, this.getClass().getClassLoader());
     }
 }