+++ src/main/java/insane96mcp/insanelib/base/config/LoadFeature.java
@@ -1,9 +1,7 @@
 package insane96mcp.insanelib.base.config;
 
-import insane96mcp.insanelib.base.Module;
-
 public @interface LoadFeature {
-    Module module();
+    String module();
     boolean enabledByDefault() default true;
     boolean canBeDisabled() default true;
 }