+++ src/main/java/insane96mcp/insanelib/module/base/feature/TagsFeature.java
@@ -16,7 +16,7 @@ import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 
-@LoadFeature(module = Module.BASE, canBeDisabled = false)
+@LoadFeature(module = "insanelib:base", canBeDisabled = false)
 @Label(name = "Tags", description = "Set and use some tags to set some mobs properties. E.g. cause fire explosion for mobs or get if a mob has been spawned from spawner.")
 public class TagsFeature extends Feature {